+++ src/yql/HISTORY.md
@@ -1,6 +1,13 @@
 YQL Change History
 ==================
 
+@VERSION@
+-----
+
+* Created a `yql-jsonp` module that requires the various JSONP modules needed. This allows us to use Conditional Loading
+to only load them when needed.
+
+
 3.8.0
 -----